# examples/ruby_rails/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rails catalog",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/products/echo": {
      "post": {
        "summary": "Echo product",
        "tags": [
          "Catalog"
        ],
        "operationId": "echoProduct",
        "parameters": [],
        "responses": {
          "200": {
            "description": "product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        }
      }
    },
...
    "/inspect/{key}": {
      "get": {
        "summary": "Inspect parsed parameters",
        "tags": [
          "Catalog"
        ],
        "operationId": "inspect",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "tags[]",
            "in": "query",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "x-trace",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          }
        }
      }
    },
    "/forms/echo": {
      "post": {
        "summary": "URL encoded form",
        "tags": [
          "Catalog"
        ],
        "operationId": "form",
        "parameters": [],
        "responses": {
          "200": {
            "description": "receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/FormFields"
              }
            }
          },
          "required": true
        }
      }
    },
    "/uploads": {
      "post": {
        "summary": "Multipart upload",
        "tags": [
          "Catalog"
        ],
        "operationId": "upload",
        "parameters": [],
        "responses": {
          "200": {
            "description": "receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadFields"
              }
            }
          },
          "required": true
        }
      }
    },
...
    "/files/echo": {
      "post": {
        "summary": "Binary upload",
        "tags": [
          "Catalog"
        ],
        "operationId": "echoBinary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "binary",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/text/{encoding}": {
      "get": {
        "summary": "Encoded text",
        "tags": [
          "Catalog"
        ],
        "operationId": "text",
        "parameters": [
          {
            "name": "encoding",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/payments/{kind}": {
      "get": {
        "summary": "Discriminated payment",
        "tags": [
          "Catalog"
        ],
        "operationId": "payment",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDocument"
                }
              }
            }
          }
        }
      }
    },
    "/payments/echo": {
      "post": {
        "summary": "Echo payment",
        "tags": [
          "Catalog"
        ],
        "operationId": "echoPayment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDocument"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDocument"
              }
            }
          }
        }
      }
    },
...
    "/customers": {
      "post": {
        "summary": "Create customer",
        "tags": [
          "Catalog"
        ],
        "operationId": "createCustomer",
        "parameters": [],
        "responses": {
          "201": {
            "description": "customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Product": {
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "customer",
          "tags",
          "attributes",
          "released",
          "updated"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "sold"
            ]
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "released": {
            "type": "string",
            "format": "date"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Receipt": {
        "type": "object",
        "required": [
          "text",
          "count",
          "enabled",
          "tags",
          "fileName",
          "contentType",
          "bytes",
          "raw"
        ],
        "properties": {
          "text": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "enabled": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "bytes": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "raw": {
            "type": "string"
          }
        }
      },
      "FormFields": {
        "type": "object",
        "required": [
          "text",
          "count",
          "enabled",
          "tags[]"
        ],
        "properties": {
          "text": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "enabled": {
            "type": "boolean"
          },
          "tags[]": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UploadFields": {
        "type": "object",
        "required": [
          "file",
          "text",
          "count",
          "enabled"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          },
          "text": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "Card": {
        "type": "object",
        "required": [
          "kind",
          "last4"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "card"
            ]
          },
          "last4": {
            "type": "string"
          }
        }
      },
      "Bank": {
        "type": "object",
        "required": [
          "kind",
          "iban"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "bank"
            ]
          },
          "iban": {
            "type": "string"
          }
        }
      },
      "PaymentDocument": {
        "type": "object",
        "required": [
          "payment"
        ],
        "properties": {
          "payment": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Card"
              },
              {
                "$ref": "#/components/schemas/Bank"
              }
            ],
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "card": "#/components/schemas/Card",
                "bank": "#/components/schemas/Bank"
              }
            }
          }
        }
      },
...
      "CustomerInput": {
        "type": "object",
        "required": [
          "name",
          "secret"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "secret": {
            "type": "string",
            "writeOnly": true
          }
        }
      },
...
    }
  }
}